Polish the docker setup
I hope this setup is clearer, and explains a bit more
into detail how to use the docker setup, and what motivates
the design choices made. I've also tried to rectify the problem
with what user owns the generated files. I'll need someone
with an true linux environment (and not just a VM) to try
it out in practise though